generate colordata in lyric fetch service, accomodate optional color data, move artwork color generation to separate extension file, call it in color generation fucntion
