+++ LyricFever/Models/NetworkFetchReturn.swift
@@ -1,0 +1,11 @@
+//
+//  NetworkFetchReturn.swift
+//  Lyric Fever
+//
+//  Created by Avi Wadhwa on 2025-08-06.
+//
+
+struct NetworkFetchReturn {
+    let lyrics: [LyricLine]
+    let colorData: Int32?
+}
